Added support for Paquette style shorthand
The parser creates the same kind of Python lists out of the shorthand
that it uses for DAIDE strings.  It then initializes an XDO object with
those lists so that we can produce both DAIDE and English glosses of
the shorthand.

Also added support for North and South coasts of Spain and St. Petersburg

Added unit tests for shorthand